replase logic from app folder to modules>header>components>Cart
